+++ backend/routes.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const { User } = require("./modals/user");
+const { User } = require("./models/user");
 
 const app = express();
